Refactoring: Slim down fat module
